Break out task text content into separate component
