+++ src/components/dashboard/taskCard/Task.js
@@ -1,6 +1,5 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui';
-import { darken } from '@theme-ui/color';
 import { useState, useContext, Fragment } from 'react';
 import TaskContext from '../../../context/TaskContext';
 import PropTypes from 'prop-types';
@@ -10,13 +9,9 @@ import DeleteTask from '../../forms/DeleteTask';
 import ViewTask from '../../ViewTask';
 import TaskCheckmark from './TaskCheckmark';
 import Card from './Card';
-import AnimatedStrikethrough from './AnimatedStrikethrough';
+import TaskCardContents from './TaskCardContents';
 import HiddenIcons from '../HiddenIcons';
 import Modal from '../Modal';
-
-const pStyles = {
-  margin: `0`,
-};
 
 const Task = ({ task }) => {
   const { editTask } = useContext(TaskContext);
@@ -25,9 +20,6 @@ const Task = ({ task }) => {
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [isViewTaskOpen, setIsViewTaskOpen] = useState(false);
-
-  const taskDueDateObj = new Date(task.due_date);
-  const dueDate = taskDueDateObj.toLocaleDateString();
 
   const toggleComplete = () => {
     // make PUT request to backend to update `is_complete` for this task
@@ -78,50 +70,11 @@ const Task = ({ task }) => {
             isChecked={isComplete}
             id={task.id}
           />
-          <div
-            sx={{
-              display: `grid`,
-              gridTemplateColumns: `repeat(2, 1fr)`,
-              gridGap: `10px`,
-            }}
-          >
-            <p
-              sx={{
-                ...pStyles,
-                display: `flex`,
-                alignItems: `center`,
-                justifyContent: `space-between`,
-                gridColumn: `1 / span 2`,
-                fontWeight: 700,
-                color: isComplete ? darken('muted', 0.4) : 'text',
-                transition: isComplete
-                  ? 'color 0.1s cubic-bezier(.55, 0, .1, 1)'
-                  : 'none',
-              }}
-            >
-              <AnimatedStrikethrough
-                stringToStrike={task.task_name}
-                isStruckOut={isComplete}
-                isNotInitial={isNotInitial}
-              />
-            </p>
-
-            <small
-              sx={{
-                ...pStyles,
-                color: isComplete ? darken('muted', 0.4) : 'text',
-                transition: isComplete
-                  ? 'color 0.1s cubic-bezier(.55, 0, .1, 1)'
-                  : 'none',
-              }}
-            >
-              <AnimatedStrikethrough
-                stringToStrike={dueDate}
-                isStruckOut={isComplete}
-                isNotInitial={isNotInitial}
-              />
-            </small>
-          </div>
+          <TaskCardContents
+            task={task}
+            isComplete={isComplete}
+            isNotInitial={isNotInitial}
+          />
         </Card>
         <button style={detailsButton} type="details" onClick={toggleViewTask}>
           Details
